# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Scene {
  id: number;
  heading?: string;
  text: string;
  visualPrompt: string;
  durationEstimate: number;
  emotionalShift?: string;
  image?: string;
}

export interface Story {
  id: string;
  title: string;
  description: string;
  mood: string;
  tags: string[];
  duration: string;
  coverImage: string;
  audioUrl?: string;
  isPremium: boolean;
  isNew: boolean;
  isSeries?: boolean;
  seriesName?: string;
  seriesEpisode?: number;
  scenes?: Scene[];
}

export interface Episode {
  id: string;
  episodeNumber: number;
  title: string;
  description: string;
  duration: string;
  isLocked: boolean;
  audioUrl?: string;
}

export interface Series {
  id: string;
  title: string;
  description: string;
  mood: string;
  coverImage: string;
  episodeCount: number;
  episodes: Episode[];
  tags: string[];
}

export type StoryBriefScenePlanItem = { [key: string]: unknown };

export interface StoryBrief {
  emotional_arc: string;
  relationship_dynamic: string;
  conflict_type: string;
  pacing_style: string;
  ending_type: string;
  sensory_palette: string[];
  point_of_view: string;
  voice_tone: string;
  scene_count: number;
  scene_plan: StoryBriefScenePlanItem[];
  recurring_motif: string;
  title_direction: string;
  image_style_direction: string;
  recommendation_tags?: string[];
  quality_target?: string;
}

export interface QcSubScores {
  emotional_depth: number;
  specificity: number;
  pacing: number;
  scene_progression: number;
  originality: number;
  sensory_detail: number;
  ending_strength: number;
}

export interface QcResult {
  passed: boolean;
  score_total: number;
  sub_scores: QcSubScores;
  issues: string[];
  rewrite_strategy?: string | null;
}

export interface GeneratedStory {
  title: string;
  description: string;
  scenes: Scene[];
}

export interface QcStoryRequest {
  brief: StoryBrief;
  story: GeneratedStory;
}

export interface RewriteStoryRequest {
  brief: StoryBrief;
  story: GeneratedStory;
  strategy: string;
}

export interface GenerateStoryRequest {
  listenerName: string;
  mood: string;
  intensity: string;
  voiceFeel: string;
  storyLength: string;
  scenarioPrompt: string;
  cinematicVisuals?: boolean;
  emotionalFocus?: boolean;
  /** When true, skips request-hash cache lookup (used for variation and continuation requests) */
  bypassCache?: boolean;
  /** Guest userId from localStorage — used to track generated story in user profile */
  userId?: string;
}

export interface GenerateStoryFromBriefRequest {
  brief: StoryBrief;
  listenerName?: string;
}

export interface GenerateImagePromptsRequest {
  brief: StoryBrief;
  story: GeneratedStory;
}

export interface SceneImagePrompt {
  sceneId: number;
  prompt: string;
}

export interface ImagePrompts {
  coverPrompt: string;
  scenePrompts: SceneImagePrompt[];
}

export interface GenerateAudioRequest {
  text: string;
  voiceFeel: string;
}

export interface GenerateAudioResponse {
  audioUrl: string;
}

export interface GenerateImagesRequest {
  coverPrompt: string;
  scenePrompts: string[];
}

export interface GenerateImagesResponse {
  cover: string;
  scenes: string[];
}

export interface FullGeneratedStory {
  id: string;
  title: string;
  description: string;
  mood: string;
  audioUrl: string;
  duration: string;
  brief?: StoryBrief;
  scenes: Scene[];
  images: GenerateImagesResponse;
  qc?: QcResult;
  recommendation_tags?: string[];
  cached?: boolean;
}

export interface SaveStoryRequest {
  userId: string;
  storyId: string;
}

export interface SaveStoryResponse {
  saved: boolean;
}

export interface UpdateProgressRequest {
  userId: string;
  storyId: string;
  audioProgressSeconds: number;
  sceneIndex: number;
}

export type UpdateTasteRequestEvent =
  (typeof UpdateTasteRequestEvent)[keyof typeof UpdateTasteRequestEvent];

export const UpdateTasteRequestEvent = {
  generated: "generated",
  replayed: "replayed",
  saved: "saved",
  completed: "completed",
} as const;

export interface UpdateTasteRequest {
  userId: string;
  mood?: string;
  intensity?: string;
  voiceFeel?: string;
  endingType?: string;
  relationshipDynamic?: string;
  event: UpdateTasteRequestEvent;
}

export type LibraryResponseSavedItem = { [key: string]: unknown };

export type LibraryResponseGeneratedItem = { [key: string]: unknown };

export interface LibraryResponse {
  saved: LibraryResponseSavedItem[];
  generated: LibraryResponseGeneratedItem[];
}

export type RecommendationsResponseForYouItem = { [key: string]: unknown };

export type RecommendationsResponseBecauseYouLikedItem = {
  [key: string]: unknown;
};

export type RecommendationsResponseContinueTheMoodItem = {
  [key: string]: unknown;
};

export interface RecommendationsResponse {
  for_you: RecommendationsResponseForYouItem[];
  because_you_liked: RecommendationsResponseBecauseYouLikedItem[];
  because_you_liked_mood?: string | null;
  continue_the_mood: RecommendationsResponseContinueTheMoodItem[];
  has_taste_profile: boolean;
}

export type GetStoriesParams = {
  mood?: string;
  search?: string;
};

export type UpdateProgress200 = {
  updated: boolean;
};

export type GetLibraryParams = {
  userId: string;
};

export type GetContinueListeningParams = {
  userId: string;
};

export type GetContinueListening200Item = { [key: string]: unknown };

export type UpdateTaste200 = {
  updated: boolean;
};