# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: stories
    description: Story library operations
  - name: series
    description: Series library operations
  - name: generate
    description: AI story generation operations
  - name: library
    description: User library, saves, progress and recommendations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /stories:
    get:
      operationId: getStories
      tags: [stories]
      summary: Get story library
      description: Returns list of curated stories
      parameters:
        - name: mood
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"

  /stories/{id}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get story by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Story detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
        "404":
          description: Not found

  /series:
    get:
      operationId: getSeries
      tags: [series]
      summary: Get series library
      responses:
        "200":
          description: List of series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Series"

  /series/{id}:
    get:
      operationId: getSeriesById
      tags: [series]
      summary: Get series by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Series detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        "404":
          description: Not found

  /plan-story:
    post:
      operationId: planStory
      tags: [generate]
      summary: Generate hidden story brief from user intake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStoryRequest"
      responses:
        "200":
          description: Story brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryBrief"

  /generate-story:
    post:
      operationId: generateStory
      tags: [generate]
      summary: Generate story text from a story brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStoryFromBriefRequest"
      responses:
        "200":
          description: Generated story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedStory"

  /qc-story:
    post:
      operationId: qcStory
      tags: [generate]
      summary: Quality-check a generated story against its brief
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcStoryRequest"
      responses:
        "200":
          description: QC evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcResult"

  /rewrite-story:
    post:
      operationId: rewriteStory
      tags: [generate]
      summary: Apply targeted rewrite to improve a specific quality dimension
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewriteStoryRequest"
      responses:
        "200":
          description: Improved story
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedStory"

  /generate-image-prompts:
    post:
      operationId: generateImagePrompts
      tags: [generate]
      summary: Generate cohesive image prompts from brief and story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateImagePromptsRequest"
      responses:
        "200":
          description: Generated image prompts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImagePrompts"

  /generate-audio:
    post:
      operationId: generateAudio
      tags: [generate]
      summary: Generate audio narration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAudioRequest"
      responses:
        "200":
          description: Generated audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateAudioResponse"

  /generate-images:
    post:
      operationId: generateImages
      tags: [generate]
      summary: Generate story images
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateImagesRequest"
      responses:
        "200":
          description: Generated images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateImagesResponse"

  /generate-full-story:
    post:
      operationId: generateFullStory
      tags: [generate]
      summary: Generate complete story with audio and images via full pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStoryRequest"
      responses:
        "200":
          description: Complete generated story with audio and images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullGeneratedStory"

  /save-story:
    post:
      operationId: saveStory
      tags: [library]
      summary: Save a story to user's library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveStoryRequest"
      responses:
        "200":
          description: Save status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveStoryResponse"
    delete:
      operationId: unsaveStory
      tags: [library]
      summary: Unsave a story from user's library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveStoryRequest"
      responses:
        "200":
          description: Save status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveStoryResponse"

  /update-progress:
    post:
      operationId: updateProgress
      tags: [library]
      summary: Save listening progress for a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgressRequest"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
                required:
                  - updated

  /library:
    get:
      operationId: getLibrary
      tags: [library]
      summary: Get user's library (saved and generated stories)
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Library content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"

  /continue-listening:
    get:
      operationId: getContinueListening
      tags: [library]
      summary: Get in-progress stories sorted by most recent
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: In-progress stories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  /update-taste:
    post:
      operationId: updateTaste
      tags: [library]
      summary: Update user taste profile after an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTasteRequest"
      responses:
        "200":
          description: Taste updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
                required:
                  - updated

  /recommendations/{userId}:
    get:
      operationId: getRecommendations
      tags: [library]
      summary: Get personalised story recommendations for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Personalised recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Story:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        mood:
          type: string
        tags:
          type: array
          items:
            type: string
        duration:
          type: string
        coverImage:
          type: string
        audioUrl:
          type: string
        isPremium:
          type: boolean
        isNew:
          type: boolean
        isSeries:
          type: boolean
        seriesName:
          type: string
        seriesEpisode:
          type: number
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
      required:
        - id
        - title
        - description
        - mood
        - tags
        - duration
        - coverImage
        - isPremium
        - isNew

    Scene:
      type: object
      properties:
        id:
          type: number
        heading:
          type: string
        text:
          type: string
        visualPrompt:
          type: string
        durationEstimate:
          type: number
        emotionalShift:
          type: string
        image:
          type: string
      required:
        - id
        - text
        - visualPrompt
        - durationEstimate

    Series:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        mood:
          type: string
        coverImage:
          type: string
        episodeCount:
          type: number
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - description
        - mood
        - coverImage
        - episodeCount
        - episodes
        - tags

    Episode:
      type: object
      properties:
        id:
          type: string
        episodeNumber:
          type: number
        title:
          type: string
        description:
          type: string
        duration:
          type: string
        isLocked:
          type: boolean
        audioUrl:
          type: string
      required:
        - id
        - episodeNumber
        - title
        - description
        - duration
        - isLocked

    StoryBrief:
      type: object
      properties:
        emotional_arc:
          type: string
        relationship_dynamic:
          type: string
        conflict_type:
          type: string
        pacing_style:
          type: string
        ending_type:
          type: string
        sensory_palette:
          type: array
          items:
            type: string
        point_of_view:
          type: string
        voice_tone:
          type: string
        scene_count:
          type: number
        scene_plan:
          type: array
          items:
            type: object
        recurring_motif:
          type: string
        title_direction:
          type: string
        image_style_direction:
          type: string
        recommendation_tags:
          type: array
          items:
            type: string
        quality_target:
          type: string
      required:
        - emotional_arc
        - relationship_dynamic
        - conflict_type
        - pacing_style
        - ending_type
        - sensory_palette
        - point_of_view
        - voice_tone
        - scene_count
        - scene_plan
        - recurring_motif
        - title_direction
        - image_style_direction

    QcSubScores:
      type: object
      properties:
        emotional_depth:
          type: number
        specificity:
          type: number
        pacing:
          type: number
        scene_progression:
          type: number
        originality:
          type: number
        sensory_detail:
          type: number
        ending_strength:
          type: number
      required:
        - emotional_depth
        - specificity
        - pacing
        - scene_progression
        - originality
        - sensory_detail
        - ending_strength

    QcResult:
      type: object
      properties:
        passed:
          type: boolean
        score_total:
          type: number
        sub_scores:
          $ref: "#/components/schemas/QcSubScores"
        issues:
          type: array
          items:
            type: string
        rewrite_strategy:
          type: string
          nullable: true
      required:
        - passed
        - score_total
        - sub_scores
        - issues

    QcStoryRequest:
      type: object
      properties:
        brief:
          $ref: "#/components/schemas/StoryBrief"
        story:
          $ref: "#/components/schemas/GeneratedStory"
      required:
        - brief
        - story

    RewriteStoryRequest:
      type: object
      properties:
        brief:
          $ref: "#/components/schemas/StoryBrief"
        story:
          $ref: "#/components/schemas/GeneratedStory"
        strategy:
          type: string
      required:
        - brief
        - story
        - strategy

    GenerateStoryRequest:
      type: object
      properties:
        listenerName:
          type: string
        mood:
          type: string
        intensity:
          type: string
        voiceFeel:
          type: string
        storyLength:
          type: string
        scenarioPrompt:
          type: string
        cinematicVisuals:
          type: boolean
        emotionalFocus:
          type: boolean
        bypassCache:
          type: boolean
          description: When true, skips request-hash cache lookup (used for variation and continuation requests)
        userId:
          type: string
          description: Guest userId from localStorage — used to track generated story in user profile
      required:
        - listenerName
        - mood
        - intensity
        - voiceFeel
        - storyLength
        - scenarioPrompt

    GenerateStoryFromBriefRequest:
      type: object
      properties:
        brief:
          $ref: "#/components/schemas/StoryBrief"
        listenerName:
          type: string
      required:
        - brief

    GeneratedStory:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
      required:
        - title
        - description
        - scenes

    GenerateImagePromptsRequest:
      type: object
      properties:
        brief:
          $ref: "#/components/schemas/StoryBrief"
        story:
          $ref: "#/components/schemas/GeneratedStory"
      required:
        - brief
        - story

    SceneImagePrompt:
      type: object
      properties:
        sceneId:
          type: number
        prompt:
          type: string
      required:
        - sceneId
        - prompt

    ImagePrompts:
      type: object
      properties:
        coverPrompt:
          type: string
        scenePrompts:
          type: array
          items:
            $ref: "#/components/schemas/SceneImagePrompt"
      required:
        - coverPrompt
        - scenePrompts

    GenerateAudioRequest:
      type: object
      properties:
        text:
          type: string
        voiceFeel:
          type: string
      required:
        - text
        - voiceFeel

    GenerateAudioResponse:
      type: object
      properties:
        audioUrl:
          type: string
      required:
        - audioUrl

    GenerateImagesRequest:
      type: object
      properties:
        coverPrompt:
          type: string
        scenePrompts:
          type: array
          items:
            type: string
      required:
        - coverPrompt
        - scenePrompts

    GenerateImagesResponse:
      type: object
      properties:
        cover:
          type: string
        scenes:
          type: array
          items:
            type: string
      required:
        - cover
        - scenes

    FullGeneratedStory:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        mood:
          type: string
        audioUrl:
          type: string
        duration:
          type: string
        brief:
          $ref: "#/components/schemas/StoryBrief"
        scenes:
          type: array
          items:
            $ref: "#/components/schemas/Scene"
        images:
          $ref: "#/components/schemas/GenerateImagesResponse"
        qc:
          $ref: "#/components/schemas/QcResult"
        recommendation_tags:
          type: array
          items:
            type: string
        cached:
          type: boolean
      required:
        - id
        - title
        - description
        - mood
        - audioUrl
        - duration
        - scenes
        - images

    SaveStoryRequest:
      type: object
      properties:
        userId:
          type: string
        storyId:
          type: string
      required:
        - userId
        - storyId

    SaveStoryResponse:
      type: object
      properties:
        saved:
          type: boolean
      required:
        - saved

    UpdateProgressRequest:
      type: object
      properties:
        userId:
          type: string
        storyId:
          type: string
        audioProgressSeconds:
          type: number
        sceneIndex:
          type: number
      required:
        - userId
        - storyId
        - audioProgressSeconds
        - sceneIndex

    UpdateTasteRequest:
      type: object
      properties:
        userId:
          type: string
        mood:
          type: string
        intensity:
          type: string
        voiceFeel:
          type: string
        endingType:
          type: string
        relationshipDynamic:
          type: string
        event:
          type: string
          enum: [generated, replayed, saved, completed]
      required:
        - userId
        - event

    LibraryResponse:
      type: object
      properties:
        saved:
          type: array
          items:
            type: object
        generated:
          type: array
          items:
            type: object
      required:
        - saved
        - generated

    RecommendationsResponse:
      type: object
      properties:
        for_you:
          type: array
          items:
            type: object
        because_you_liked:
          type: array
          items:
            type: object
        because_you_liked_mood:
          type: string
          nullable: true
        continue_the_mood:
          type: array
          items:
            type: object
        has_taste_profile:
          type: boolean
      required:
        - for_you
        - because_you_liked
        - continue_the_mood
        - has_taste_profile