# lib/api-zod/src/generated/types/generateStoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateStoryRequest {
  listenerName: string;
  mood: string;
  intensity: string;
  voiceFeel: string;
  storyLength: string;
  scenarioPrompt: string;
  cinematicVisuals?: boolean;
  emotionalFocus?: boolean;
  /** When true, skips request-hash cache lookup (used for variation and continuation requests) */
  bypassCache?: boolean;
  /** Guest userId from localStorage — used to track generated story in user profile */
  userId?: string;
}